+++ README.md
@@ -1,6 +1,6 @@
 SOLID
 
-SRP - разделил класс на два класса, TaskCreater и TaskService.
+SRP - разделил класс на два класса, TaskCreater и service.TaskService.
 Первый описывает объект класса, второй производит операции
 над обектами первого класса. Появилась некоторая модульность.
