+++ src/main/java/service/Service.java
@@ -1,4 +1,6 @@
-import Entities.Task;
+package service;
+
+import model.Task;
 
 import java.util.Map;
